% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{spatial_grid}
\alias{spatial_grid}
\title{a uniform two dimensional grid of lattice points}
\usage{
spatial_grid(lat, long)
}
\arguments{
\item{lat}{numeric vector}

\item{long}{numeric vector}
}
\value{
returns a dataframe of lattice points
}
\description{
a uniform two dimensional grid of lattice points
}
\examples{
spatial_grid(c(1:10),c(1:10))
}
