% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_POP.R
\name{spatialPOP}
\alias{spatialPOP}
\title{Generation of Spatial Data  with Spatial Coordinates and Spatially Varying Model Parameters}
\usage{
spatialPOP(N, m, xlat, ylong)
}
\arguments{
\item{N}{integer; population size i.e. N= (m * m)}

\item{m}{integer}

\item{xlat}{numeric vector}

\item{ylong}{numeric vector}
}
\value{
returns a dataframe of spatially generated population consist of simulated response variable (i.e. Y) along with their spatial coordinates, spatially varying model parameters and one explanatory variable (i.e. X)
}
\description{
Generation of Spatial Data  with Spatial Coordinates and Spatially Varying Model Parameters
}
\examples{
lattice_points<-spatial_grid(c(1:10),c(1:10))
spatial_data<-spatialPOP(100,10,c(1:10),c(1:10))
}
\references{
1. Leung, Y., Mei, C. L. and Zhang, W. X. (2000). Statistical tests for spatial non-stationarity based on the geographically
               weighted regression model. Environment and Planning A, 32(1), 9-32.


 2. Chao, Liu., Chuanhua, Wei. and Yunan, Su. (2018). Geographically weighted regression model-assisted estimation in survey sampling. Journal of Nonparametric Statistics. <DOI:10.1080/10485252.2018.1499907>.
}
